Fix confusion with `this` shadowing
The article previously described `this` as a dynamically scoped name, but `this` in JavaScript is a lexical name (that is, its meaning is defined statically in the source code). The problem is really that every (regular) function defines its own (hidden) `this` parameter, so they shadow the names defined in the outer functions.

The same thing would happen if one decided to name their first parameter in every function `a`.